feat(transactions): add useTransactionsData hook for balance and transactions queries
